1.修改赞助返回类型为bigdecimal

+++ carloan-service/carloan-service-admin/src/main/java/com/carloan/service/admin/mayi/vo/ZanzhuVo.java
@@ -3,6 +3,7 @@ package com.carloan.service.admin.mayi.vo;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 @Getter
@@ -10,7 +11,7 @@ import java.util.Date;
 public class ZanzhuVo {
     public int id;
     public String name;
-    public int money;
+    public BigDecimal money;
     public String remark;
     public String qq;
     public Date createtime;
